+++ apps/src/aichat/views/modelCustomization/SetupCustomization.tsx
@@ -6,7 +6,7 @@ import React, {useState, useMemo} from 'react';
 import {useSelector} from 'react-redux';
 
 import {ModelDescription} from '@cdo/apps/aichat/types';
-import {isReadOnlyWorkspace} from '@cdo/apps/lab2/lab2Redux';
+import {isReadOnlyWorkspace} from '@cdo/apps/lab2/redux/lab2ReduxSelectors';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 
 import {modelDescriptions} from '../../constants';
