+++ apps/src/lab2/lab2Redux.ts
@@ -23,10 +23,8 @@ import {
   setUserRoleInCourse,
   CourseRoles,
 } from '@cdo/apps/templates/currentUserRedux';
-import {LevelStatus} from '@cdo/generated-scripts/sharedConstants';
 
 import {setLevel} from '../aiTutor/redux/aiTutorRedux';
-import {getCurrentLevel} from '../code-studio/progressReduxSelectors';
 import {
   setProjectUpdatedAt,
   setProjectUpdatedError,
@@ -47,6 +45,7 @@ import ProjectManager from './projects/ProjectManager';
 import ProjectManagerFactory from './projects/ProjectManagerFactory';
 import {getPredictResponse} from './projects/userLevelsApi';
 import {setProjectTooLarge} from './redux/lab2ProjectRedux';
+import {isReadOnlyWorkspace} from './redux/lab2ReduxSelectors';
 import {LevelPropertiesValidator} from './responseValidators';
 import {
   Channel,
@@ -296,42 +295,6 @@ export const isLabLoading = (state: {lab: LabState}) =>
   state.lab.isLoadingProjectOrLevel ||
   state.lab.isLoading ||
   state.lab.isLoadingTheme;
-
-// This may depend on more factors, such as share.
-export const isReadOnlyWorkspace = (state: RootState) => {
-  const isEditMode = !!getAppOptionsEditBlocks();
-  const isEditingExemplar = getAppOptionsEditingExemplar();
-  const isViewingExemplar = getAppOptionsViewingExemplar();
-  const isWidgetView = !!state.lab.levelProperties?.widgetView;
-
-  // Exemplar and block edit modes do not have a channel.
-  if (isEditMode || isEditingExemplar) {
-    return false;
-  } else if (isViewingExemplar) {
-    return true;
-  }
-  // Otherwise, we are in read only mode if we are not the owner of the channel,
-  // the level is frozen, the level is a read only predict level, the level has been submitted.
-  // or this is a lab that should be read only while running and the code is currently running.
-  const isOwner = state.lab.channel?.isOwner;
-  const isFrozen = !!state.lab.channel?.frozen;
-  const readonlyPredictLevel = isReadonlyPredictLevel(state);
-  const hasSubmitted = getCurrentLevel(state)?.status === LevelStatus.submitted;
-  const isViewingOldVersion = state.lab2Project.viewingOldVersion;
-  const isRunningAndReadonly =
-    (state.lab2System.isRunning || state.lab2System.isValidating) &&
-    shouldBeReadonlyWhileRunning(state);
-
-  return (
-    !isOwner ||
-    isFrozen ||
-    readonlyPredictLevel ||
-    hasSubmitted ||
-    isRunningAndReadonly ||
-    isViewingOldVersion ||
-    isWidgetView
-  );
-};
 
 // If there is an error present on the page.
 export const hasPageError = (state: {lab: LabState}) => {
@@ -588,31 +551,6 @@ async function cleanUpProjectManager() {
   // saves from the existing project manager.
   await existingProjectManager?.cleanUp();
   Lab2Registry.getInstance().clearProjectManager();
-}
-
-// Returns if the current state represents a predict level that should be read only.
-// If the predict level code is not editable after submit or the user has not submitted a response,
-// the predict level is read only.
-function isReadonlyPredictLevel(state: RootState) {
-  const isPredictLevel =
-    state.lab.levelProperties?.predictSettings?.isPredictLevel || false;
-  let isReadonlyPredictLevel = isPredictLevel;
-  if (isPredictLevel) {
-    const isEditableAfterSubmit =
-      state.lab.levelProperties?.predictSettings?.codeEditableAfterSubmit ||
-      false;
-    const hasSubmittedPredictResponse = state.predictLevel.hasSubmittedResponse;
-    // If the predict level code is not editable after submit or the user has not submitted a response,
-    // the predict level is read only.
-    isReadonlyPredictLevel =
-      !isEditableAfterSubmit || !hasSubmittedPredictResponse;
-  }
-  return isReadonlyPredictLevel;
-}
-
-// Currently only Python Lab disables editing while code is running.
-function shouldBeReadonlyWhileRunning(state: RootState) {
-  return state.lab.levelProperties?.appName === 'pythonlab';
 }
 
 // This is an action that other reducers (specifically predictLevelRedux) can respond to.
